updated root version

+++ modules/docker/docker/install/start.sh
@@ -12,10 +12,29 @@ if [ ! -f docker-compose.yml ]; then
 fi
 
 # Set local environment
-if [ -f setLocalEnv.sh ]; then
-    . ./setLocalEnv.sh
+if [ ! -f setLocalEnv.sh ]; then
+    cp setLocalEnvSample.sh setLocalEnv.sh
+    chmod +x setLocalEnv.sh
+fi
+. ./setLocalEnv.sh
+
+# We need the deployment environment
+if [ -z "$DEPLOY_ENV" ]; then
+    echo "Error: No DEPLOY_ENV environment variable set. Please check the contents of setLocalEnv.sh"
+    exit 1
 fi
 
+# We need the configuration directory
+if [ ! -f "$SCRIPT_DIR/config/Dockerfile" ]; then
+    echo "Error: $SCRIPT_DIR/config/Dockerfile does not exist. Unable to create the configuration container"
+    exit 1
+fi
+
+# Create the configuration container
+docker build -t docker.appiness.mobi/apptvate-$DEPLOY_ENV-web-config:latest $SCRIPT_DIR/config
+docker push docker.appiness.mobi/apptvate-$DEPLOY_ENV-web-config:latest
+
+# Restart
 docker-compose down
 docker-compose pull
 docker-compose up -d --force-recreate web
